% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/beer.R
\docType{data}
\name{beer}
\alias{beer}
\title{Monthly beer production in Australia}
\format{
477 x 2 dataframe

\describe{
\item{column1}{Data}
\item{column2}{prodution}
}
}
\usage{
beer
}
\description{
Monthly beer production in Australia
}
\keyword{datasets}
