% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Carsales.R
\docType{data}
\name{Carsales}
\alias{Carsales}
\title{Top Cars Sales}
\format{
3148 x 12 dataframe

\describe{
\item{index}{id}
\item{column1}{salesperson}
\item{column2}{city}
}
}
\usage{
Carsales
}
\description{
Top Cars Sales
}
\keyword{datasets}
