% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/License.R
\docType{data}
\name{License}
\alias{License}
\title{Shanghai License plate deal price}
\format{
190 x 6 dataframe

\describe{
\item{index}{id}
\item{column1}{Date}
\item{column2}{avg_deal_price （$)}
\item{column3}{lowest_deal_price ($)}
\item{column4}{num_bidder}
\item{column5}{num_plates}
}
}
\usage{
License
}
\description{
Shanghai License plate deal price
}
\keyword{datasets}
