% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/License_price.R
\docType{data}
\name{License_price}
\alias{License_price}
\title{Shanghai License Plate deal price data}
\format{
190 x 6 dataframe

\describe{
\item{index}{ID}
\item{column1}{Date}
\item{column2}{avg_deal_price （$)}
\item{column3}{lowest_deal_price ($)}
\item{column4}{num_bidder}
\item{column5}{num_plates}
}
}
\source{
https://www.kaggle.com/bazingasu/shanghai-license-plate-bidding-price-prediction
}
\usage{
License_price
}
\description{
Shanghai License Plate deal price data
}
\keyword{datasets}
