% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Plate_price.R
\docType{data}
\name{Plate_price}
\alias{Plate_price}
\title{Shanghai license plate deal price}
\format{
166 x 3 dataframe
}
\usage{
Plate_price
}
\description{
Shanghai license plate deal price
}
\keyword{datasets}
