% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/License plate price.R
\docType{data}
\name{Shanghai_license_plate_price}
\alias{Shanghai_license_plate_price}
\title{Shanghai_license_plate_deal_price}
\format{
166 x 4 dataframe
}
\usage{
Shanghai_license_plate_price
}
\description{
Shanghai_license_plate_deal_price
}
\keyword{datasets}
