% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/beer.R
\docType{data}
\name{beer}
\alias{beer}
\title{Monthly beer production in australia}
\format{
477 x 2 dataframe

\describe{
\item{column1}{Date}
\item{column2}{Monthly beer production}
}
}
\usage{
beer
}
\description{
Monthly beer production in australia
}
\keyword{datasets}
